Reviewed Auth and Authorization
- Updated server (done)
- Client being updated.

+++ server/classes/user.js
@@ -1,6 +1,8 @@
 const pool = require("../db");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
+const ROLES_LIST = require("../config/roles_list");
+require("dotenv").config();
 
 class User {
   static async getUserType(username) {
@@ -25,29 +27,48 @@ class User {
     }
   } //getUsernames
 
-  static async signUp(userName, password, userType) {
-    const salt = bcrypt.genSaltSync(10);
-    const hashedPassword = bcrypt.hashSync(password, salt);
+  static async signUp(req, res) {
+    const { userName, password, userType } = req.body;
+    //check if the username and password are filled in
+    if (!userName || !password)
+      return res
+        .status(400)
+        .json({ message: "Username and password are required." });
+
+    //check if the user is duplicated
+    const users = await this.getUsernames();
+    const duplicate = users.find((user) => user.username === userName);
+    if (duplicate)
+      return res.status(409).json({ message: "User already exists" }); //conflict
+
     try {
-      //insert the user data into users table
-      const signUp = await pool.query(
+      //encript password
+      const salt = bcrypt.genSaltSync(10);
+      const hashedPassword = bcrypt.hashSync(password, salt);
+
+      //store the user data into users table
+      const dbResponse = await pool.query(
         `INSERT INTO users (username, hashed_password, user_type) VALUES ($1, $2, $3)`,
         [userName, hashedPassword, userType]
       );
 
       //send the reponse to client
-      return { userName };
+      return res.status(201).json({ userName });
     } catch (error) {
       console.error(error);
-
-      //if error send the detail to client
-      if (error) {
-        return { detail: error.detail };
-      }
+      return res.status(500).json({ message: error.message });
     }
   } //signUp
 
-  static async login(userName, password, res) {
+  static async login(req, res) {
+    const { userName, password } = req.body;
+
+    //check if the username and password are filled in
+    if (!userName || !password)
+      return res
+        .status(400)
+        .json({ message: "Username and password are required." });
+
     try {
       //Find the user in the database
       const users = await pool.query(
@@ -56,7 +77,8 @@ class User {
       );
 
       //if user not found
-      if (!users.rows.length) return { detail: "User does not exist." };
+      if (!users.rows.length)
+        return res.status(401).json({ message: "User does not exist." });
 
       //check user password with the password in database
       const success = await bcrypt.compare(
@@ -64,37 +86,79 @@ class User {
         users.rows[0].hashed_password
       );
 
-      //get the userType to include in the token
-      const userType = users.rows[0].user_type;
-
-      //create a token with the user deitalis
-      const token = jwt.sign({ userName, userType }, "statesecret", {
-        expiresIn: "1hr",
-      });
-
       if (success) {
-        return {
+        //get the userType to include in the token
+        const userType = users.rows[0].user_type;
+
+        //get user type code
+        const userTypeCode = ROLES_LIST[userType];
+
+        //create access token with the user deitalis
+        const accessToken = jwt.sign(
+          { username: userName, role: userTypeCode },
+          process.env.ACCESS_TOKEN_SECRET,
+          {
+            expiresIn: "15m",
+          }
+        );
+
+        //create refresh token with the user deitalis
+        const refreshToken = jwt.sign(
+          { userName },
+          process.env.REFRESH_TOKEN_SECRET,
+          {
+            expiresIn: "10hr",
+          }
+        );
+
+        //update the user in DB with the refresh token
+        const updatedUser = await pool.query(
+          "UPDATE users SET refresh_token = $1 WHERE username = $2",
+          [refreshToken, userName]
+        );
+
+        //check if the user was updated
+        if (!updatedUser?.rowCount)
+          return res
+            .status(500)
+            .json({ message: "Failed to update refresh token" });
+
+        //send the refresh token to client within a httpCookie
+        res.cookie("jwt", refreshToken, {
+          httpOnly: true,
+          sameSite: "None",
+          secure: true,
+          maxAge: 10 * 60 * 60 * 1000,
+        });
+
+        return res.json({
           username: users.rows[0].username,
-          token,
-        };
+          accessToken,
+          role: userTypeCode,
+        });
       } else {
-        return { detail: "Login failed" };
+        return res.status(401).json({ message: "Login failed" });
       }
     } catch (error) {
-      res.status(503).send("Database unavailable");
       console.error(error);
+      return res.status(500).json({ message: error.message });
     }
   } //login
 
-  static async resetPassword(userName, currentPassword, newPassword) {
+  static async resetPassword(req, res) {
+    const { userName, currentPassword, newPassword } = req.body;
+
+    if (!userName || !currentPassword || !newPassword)
+      return res
+        .status(400)
+        .json({ message: "Username and Password required." });
+
     try {
       //select the user by his username
       const users = await pool.query(
         "SELECT * FROM users WHERE username = $1",
         [userName]
       );
-
-      if (!users.rows.length) return { detail: "User does not exist." };
 
       //validate the current password
       const passwordMatch = await bcrypt.compare(
@@ -112,20 +176,27 @@ class User {
         );
 
         if (passwordUpdated) {
-          return "Password updated";
+          return res.json("Password updated");
         } else {
-          return { detail: "Error updating password" };
+          return res.status(503).json({ message: "Error updating password" });
         }
       } else {
-        return { detail: "Password does not match" };
+        return res.status(400).json({ message: "Password does not match" });
       }
     } catch (err) {
       console.error(err);
-      if (err) return { detail: err.detail };
+      return res.status(500).json({ message: err.message });
     }
   } //resetPassword
 
-  static async resetUserPass(userName, newPassword) {
+  static async resetUserPass(req, res) {
+    const { userName, newPassword } = req.body;
+
+    if (!userName || !newPassword)
+      return res
+        .status(400)
+        .json({ message: "Username and Password required." });
+
     try {
       //select the user by his username
       const users = await pool.query(
@@ -133,7 +204,8 @@ class User {
         [userName]
       );
 
-      if (!users.rows.length) return { detail: "User does not exist." };
+      if (!users.rows.length)
+        return res.status(400).json({ message: "User does not exist." });
 
       //update password
       const salt = bcrypt.genSaltSync(10);
@@ -144,26 +216,32 @@ class User {
       );
 
       if (passwordUpdated) {
-        return "Password updated";
+        return res.json("Password updated");
       } else {
-        return { detail: "Error updating password" };
+        return res.status(503).json({ message: "Error updating password" });
       }
     } catch (err) {
       console.error(err);
-      if (err) return { detail: err.detail };
+      return res.status(500).json({ message: err.message });
     }
   } //resetUserPass
 
-  static async deleteUser(userName) {
+  static async deleteUser(req, res) {
+    const { username } = req.params;
+
+    if (!username)
+      return res.status(400).json({ message: "Username is required" });
+
     try {
       const deletedUser = await pool.query(
         "DELETE from users WHERE username = $1",
-        [userName]
+        [username]
       );
-      return deletedUser;
+      console.log(deletedUser);
+      return res.json(deletedUser);
     } catch (err) {
       console.error(err);
-      if (err) return { detail: err.detail };
+      return res.status(500).json({ message: err.message });
     }
   } //deleteUser
 }
